BasicParserContext takes DecimalNumberContext instead of literals.
- renamed DecimalNumberContextParserContext to BasicParserContext.
- Introduced DecimalNumberContexts & DecimalNumberContextTestCase.
