+++ src/test/java/walkingkooka/text/cursor/parser/BigIntegerParserTest.java
@@ -18,6 +18,7 @@ package walkingkooka.text.cursor.parser;
 
 import org.junit.Test;
 import walkingkooka.Cast;
+import walkingkooka.DecimalNumberContexts;
 import walkingkooka.text.cursor.TextCursor;
 import walkingkooka.text.cursor.TextCursors;
 
@@ -121,7 +122,7 @@ public class BigIntegerParserTest extends ParserTemplateTestCase<BigIntegerParse
 
     private TextCursor parseAndCheck3(final String text, final int value) {
         return this.parseAndCheck(this.createParser(),
-                ParserContexts.basic('!', 'X', 'M', 'P'),
+                ParserContexts.basic(DecimalNumberContexts.basic('!', 'X', 'M', 'P')),
                 text,
                 ParserTokens.bigInteger(BigInteger.valueOf(value), text),
                 text,
@@ -145,7 +146,7 @@ public class BigIntegerParserTest extends ParserTemplateTestCase<BigIntegerParse
 
     @Override
     protected ParserContext createContext() {
-        return ParserContexts.basic('.', 'E', '-', '+');
+        return ParserContexts.basic(DecimalNumberContexts.basic('.', 'E', '-', '+'));
     }
 
     private TextCursor parseAndCheck2(final String in, final long value, final String text, final String textAfter){
