+++ src/test/java/walkingkooka/text/cursor/parser/DoubleParserTest.java
@@ -19,6 +19,7 @@ package walkingkooka.text.cursor.parser;
 
 import org.junit.Test;
 import walkingkooka.Cast;
+import walkingkooka.DecimalNumberContexts;
 import walkingkooka.text.cursor.TextCursor;
 
 import static org.junit.Assert.assertEquals;
@@ -462,7 +463,7 @@ public final class DoubleParserTest extends ParserTemplateTestCase<DoubleParser<
 
     private TextCursor parseAndCheck3(final String text, final double value) {
         return this.parseAndCheck(this.createParser(),
-                ParserContexts.basic('!', 'X', 'M', 'P'),
+                ParserContexts.basic(DecimalNumberContexts.basic('!', 'X', 'M', 'P')),
                 text,
                 ParserTokens.doubleParserToken(value, text),
                 text,
@@ -476,7 +477,7 @@ public final class DoubleParserTest extends ParserTemplateTestCase<DoubleParser<
 
     @Override
     protected ParserContext createContext() {
-        return ParserContexts.basic('.', 'E', '-', '+');
+        return ParserContexts.basic(DecimalNumberContexts.basic('.', 'E', '-', '+'));
     }
 
     private TextCursor parseAndCheck2(final String in, final double value){
