+++ src/test/java/walkingkooka/text/cursor/parser/LongParserTest.java
@@ -19,6 +19,7 @@ package walkingkooka.text.cursor.parser;
 
 import org.junit.Test;
 import walkingkooka.Cast;
+import walkingkooka.DecimalNumberContexts;
 import walkingkooka.text.cursor.TextCursor;
 import walkingkooka.text.cursor.TextCursors;
 
@@ -187,7 +188,7 @@ public class LongParserTest extends ParserTemplateTestCase<LongParser<ParserCont
 
     private TextCursor parseAndCheck3(final String text, final long value) {
         return this.parseAndCheck(this.createParser(),
-                ParserContexts.basic('!', 'X', 'M', 'P'),
+                ParserContexts.basic(DecimalNumberContexts.basic('!', 'X', 'M', 'P')),
                 text,
                 ParserTokens.longParserToken(value, text),
                 text,
@@ -211,7 +212,7 @@ public class LongParserTest extends ParserTemplateTestCase<LongParser<ParserCont
 
     @Override
     protected ParserContext createContext() {
-        return ParserContexts.basic('.', 'E', '-', '+');
+        return ParserContexts.basic(DecimalNumberContexts.basic('.', 'E', '-', '+'));
     }
 
     private TextCursor parseAndCheck2(final String in, final long value, final String text, final String textAfter){
