+++ src/test/java/walkingkooka/tree/expression/ExpressionNodeEvaluationTest.java
@@ -19,13 +19,13 @@
 package walkingkooka.tree.expression;
 
 import org.junit.Test;
+import walkingkooka.DecimalNumberContexts;
 import walkingkooka.test.TestCase;
 import walkingkooka.text.CharSequences;
 import walkingkooka.text.cursor.TextCursor;
 import walkingkooka.text.cursor.TextCursorSavePoint;
 import walkingkooka.text.cursor.TextCursors;
 import walkingkooka.text.cursor.parser.Parser;
-import walkingkooka.text.cursor.parser.ParserContexts;
 import walkingkooka.text.cursor.parser.Parsers;
 import walkingkooka.text.cursor.parser.spreadsheet.SpreadsheetParserContext;
 import walkingkooka.text.cursor.parser.spreadsheet.SpreadsheetParserContexts;
@@ -200,7 +200,7 @@ public class ExpressionNodeEvaluationTest extends TestCase {
     private SpreadsheetParserToken parse(final String parse){
         final TextCursor cursor = TextCursors.charSequence(parse);
         final Optional<SpreadsheetParserToken> spreadsheetFormula = this.createParser().parse(cursor,
-                SpreadsheetParserContexts.basic(ParserContexts.basic('.', 'E', '-', '+')));
+                SpreadsheetParserContexts.basic(DecimalNumberContexts.basic('.', 'E', '-', '+')));
         if(!spreadsheetFormula.isPresent()){
             fail("Parser failed to parse " + CharSequences.quoteAndEscape(parse));
         }
